fix: Model errors. :D

+++ Journal-Limpet.Shared/Models/Journal/UserJournal.cs
@@ -22,13 +22,13 @@ namespace Journal_Limpet.Shared.Models.Journal
 
         public UserJournal(DataRow row)
         {
-            JournalId = row.Field<int>("journal_id");
+            JournalId = row.Field<long>("journal_id");
             UserIdentifier = row.Field<Guid>("user_identifier");
             Created = new DateTimeOffset(row.Field<DateTime>("created"), TimeSpan.Zero);
             JournalDate = new DateTimeOffset(row.Field<DateTime>("journal_date"), TimeSpan.Zero);
             S3Path = row["s3_path"].ToString();
             LastProcessedLine = row["last_processed_line"].ToString();
-            LastProcessedLineNumber = row.Field<int?>("last_processed_line_number");
+            LastProcessedLineNumber = row.Field<long?>("last_processed_line_number");
             CompleteEntry = row.Field<bool>("complete_entry");
             LastUpdate = !row.IsNull("last_update") ? new DateTimeOffset(row.Field<DateTime>("last_update"), TimeSpan.Zero) as DateTimeOffset? : null;
